requirements.txt
geopy==2.4.1
paho-mqtt==1.5.0
pytest==8.3.4